Plot App: Created Simple Plotting App
Created a cmake file for the project to include libraries such as iio and matplotlib_cpp
Created the main.cpp where it connects to the device through the IP address of the Cora board.
The script reads values in a buffer, translates the values to 1 value per axis.
These values are later plotted using the matplotlib library.

Signed-off-by: Gadalean Tudor <tudi.gadalean@gmail.com>