Created Python Key Simulator
Created the app.py file that simulates key presses using the
accellerometer data.
Creates two threads, one for the GUI and one for reading data.
The data acquisition is done the same as in the main.cpp file.
The data is transformed into roll and pitch values, which are
transformed in 4 coordinates: front, left, back and right.
The directions are mapped to WASD keys.
GUI presents 4 squares colored based on the inclination of the
board.

Signed-off-by: Gadalean Tudor <tudi.gadalean@gmail.com>